# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Interview Guide Platform API",
    "description": "High-performance Rust Axum API for Interview Guide & Technical Knowledge Platform",
    "contact": {
      "name": "Platform Engineering",
      "email": "eng@interviewguide.internal"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0"
  },
  "paths": {
    "/api/audit/events": {
      "get": {
        "tags": [
          "Enterprise"
        ],
        "summary": "GET /api/audit/events",
        "operationId": "list_events",
        "responses": {
          "200": {
            "description": "List recent platform audit ledger events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Admin access required"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in with email and password.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in; refresh cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Current signed-in user's profile.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange the HTTP-only refresh cookie for a new access token (rotates the refresh token).",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "New access token; refresh cookie rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create an account and start a session.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created; refresh cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.error.ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/webauthn/login/finish": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/login/finish",
        "operationId": "login_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyLoginFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passkey authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/auth/webauthn/login/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/login/start",
        "operationId": "login_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyLoginStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passkey login challenge generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyLoginStartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/webauthn/register/finish": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/register/finish",
        "operationId": "register_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegisterFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passkey credential registered successfully"
          },
          "400": {
            "description": "Verification failed"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/webauthn/register/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "POST /api/auth/webauthn/register/start",
        "operationId": "register_start",
        "responses": {
          "200": {
            "description": "Passkey registration challenge generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyRegisterStartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/bookmarks": {
      "get": {
        "tags": [
          "Bookmarks"
        ],
        "operationId": "get_bookmarks",
        "responses": {
          "200": {
            "description": "List all user bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookmarkItem"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Bookmarks"
        ],
        "operationId": "toggle_bookmark",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bookmark removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkItem"
                }
              }
            }
          },
          "201": {
            "description": "Bookmark created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/categories/{category_id}/questions": {
      "get": {
        "tags": [
          "Questions"
        ],
        "operationId": "get_questions_by_category",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty: Beginner, Intermediate, Advanced",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of questions to return (default: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of questions for the specified category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Question"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
    "/api/categories/{category_id}/questions/{q_num}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "operationId": "get_single_question",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q_num",
            "in": "path",
            "description": "Question number (1-100)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single question details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "404": {
            "description": "Question not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventItem": {
        "type": "object",
        "required": [
          "sequence_id",
          "event_type",
          "payload_json",
          "prev_hash",
          "curr_hash",
          "created_at"
        ],
        "properties": {
          "actor_id": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "curr_hash": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "payload_json": {},
          "prev_hash": {
            "type": "string"
          },
          "sequence_id": {
            "type": "integer",
            "format": "int64"
          },
          "target_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}